backend/app/requirements.txt
#py -m pip install -r requirements.txt

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-services-api
backtrader
django
djangorestframework
requests
TA-Lib
scipy
pandas